requirements.txt
python-telegram-bot[job-queue,rate-limiter]==20.7
python-dotenv
aiosqlite